// frontend/src/client/types.gen.ts
export type CompanyType = "production" | "negoce" | "service";

export type PartnerType = "entreprise" | "prestataire_logistique";

export type ActivitySector =
  | "agroalimentaire"
  | "construction_btp"
  | "industriel_manufacturier"
  | "chimique_petrochimique"
  | "agricole_rural"
  | "logistique_messagerie"
  | "medical_parapharmaceutique"
  | "hygiene_dechets_environnement"
  | "energie_ressources_naturelles"
  | "logistique_speciale"
  | "autre";
// ...
export type CompanyCreate = {
  company_name: string;
  nis: string;
  nif: string;
  headquarters_address: string;
  company_type: CompanyType;
  activity_sector: ActivitySector;
  sector_specification?: string | null;
  partner_type: PartnerType;
  legal_representative_name: string;
  legal_representative_contact: string;
  logo_url?: string | null;
};